for J2735_202409 updates

+++ J2735Common/asnSpec.py
@@ -19,5 +19,31 @@ asn1_files = [
 j2735_spec = asn1tools.compile_files(asn1_files, 'uper')
 
     # Save the compiled ASN.1 specification to a file
-with open('j2735_spec_2.pkl', 'wb') as f:
-    pickle.dump(j2735_spec, f)
+if not os.path.exists('./pkl/pklj2735_spec_2.pkl'):
+    with open('./pkl/j2735_spec_2.pkl', 'wb') as f:
+        pickle.dump(j2735_spec, f)
+
+# updae for 2024 version
+asn_path = './V2X_ASN1_Module_Collection_2024'
+asn1_files = [
+    os.path.join(asn_path, 'J2735-ITIS-2024-rel-v1.1.asn'),
+    os.path.join(asn_path, 'J2735-Common-2024-rel-v1.1.2.asn'),
+    os.path.join(asn_path, 'J2735-MapData-2024-rel-v1.1.asn'),
+    os.path.join(asn_path, 'J2735-ProbeVehicleData-2024-rel-v1.1.asn'),
+    os.path.join(asn_path, 'J2945-3-RoadWeatherMessage-2024-rel-v2.1.asn'),
+    os.path.join(asn_path, 'J2945-C-ProbeDataConfig-2024-rel-v1.1.asn'),
+    os.path.join(asn_path, 'J2945-C-ProbeDataReport-2024-rel-v1.1.asn'),
+    os.path.join(asn_path, 'J2735-BasicSafetyMessage-2024-rel-v1.1.2.asn'),
+    os.path.join(asn_path, 'J2735-NTCIP-2024-rel-v1.1.asn'),
+    os.path.join(asn_path, 'J2735-MessageFrame-2024-rel-v1.1.1.asn'),
+    os.path.join(asn_path, 'J2735-SPAT-2024-rel-v1.1.1.asn'),
+    os.path.join(asn_path, 'J2735-TravelerInformation-2024-rel-v1.1.2.asn'),
+]
+
+# Load the ASN.1 specification for J2735
+j2735_spec = asn1tools.compile_files(asn1_files, 'uper')
+
+if not os.path.exists('./pkl/j2735_spec_2024.pkl'):
+    # Save the compiled ASN.1 specification to a file
+    with open('./pkl/j2735_spec_2024.pkl', 'wb') as f:
+        pickle.dump(j2735_spec, f)
